Parse embeddings as floats before calculating distances, fixed error on euclidean distance calculation

+++ analysis/metrics/MRR_and_NDCG.py
@@ -22,8 +22,9 @@ class Metric:
         results_dict = {}
 
         for distance_metric in distance_metrics:
+            
             # Get the distances between each embedding
-            mutual_distances = spatial.distance.squareform(spatial.distance.pdist(embeddings, metric=distance_metric))
+            mutual_distances = spatial.distance.squareform(spatial.distance.pdist(embeddings.astype(np.single), metric=distance_metric))
 
             # Make the nan and inf values (which only appear in cosine and JS distances) 2, higher than the maximum for these metrics
             mutual_distances[np.isnan(mutual_distances) | np.isinf(mutual_distances)] = 2
